+++ NEWS.md
@@ -1,13 +1,14 @@
 # pestr 0.8.0
 
-## Function updates
+## New functions and updates
 
-* added "devel" on Ubuntu 18.04 to github actions.
 * `eppo_tabletool` functions now can use "raw" eppocodes instead of result of `eppo_names_tables` function result to query **EPPO Data Services** and **EPPO Global Database**. New functionality works with `raw_eppocodes` and 'use_raw_codes` parameters.
+* `msg_helpers` small wrapper over messages to reduce redundancy in code.
 
 ## Fixed issues and smaller functionalities
 
 * updated documentation of functions, including examples, typos, etc.
+* added "devel" on Ubuntu 18.04 to github actions.
 
 # pestr 0.7.4